Rename to better name update_description
