+++ src/python/ensembl/io/genomio/annotation/update_description.py
